move quote error analytics fn to a new file
