+++ src/__swaps__/screens/Swap/hooks/useSwapInputsController.ts
@@ -40,52 +40,9 @@ import { triggerHaptics } from 'react-native-turbo-haptics';
 import { useDebouncedCallback } from 'use-debounce';
 import { NavigationSteps } from './useSwapNavigation';
 import { deepEqualWorklet } from '@/worklets/comparisons';
-
-import { EventProperties } from '@/analytics/event';
+import { analyticsTrackQuoteFailed } from './analyticsTrackQuoteFailed';
 
 const REMOTE_CONFIG = getRemoteConfig();
-
-const analyticsTrack: typeof analyticsV2.track = (...args) => analyticsV2.track(...args);
-let lastParams: EventProperties[typeof analyticsV2.event.swapsQuoteFailed];
-function analyticsTrackQuoteFailed(
-  quote: QuoteError | null,
-  {
-    inputAsset,
-    outputAsset,
-    inputAmount,
-    outputAmount,
-  }: {
-    inputAsset: ExtendedAnimatedAssetWithColors | null;
-    outputAsset: ExtendedAnimatedAssetWithColors | null;
-    inputAmount: number | undefined;
-    outputAmount: number | undefined;
-  }
-) {
-  'worklet';
-  // we are tracking 'Insufficient funds' 'Out of gas' 'No routes found' and 'No quotes found'
-  if (!quote || !inputAsset || !outputAsset || !inputAmount) return;
-
-  if (
-    lastParams &&
-    quote.error_code === lastParams.error_code &&
-    inputAmount === lastParams.inputAmount &&
-    outputAmount === lastParams.outputAmount &&
-    inputAsset.address === lastParams.inputAsset.address &&
-    outputAsset.address === lastParams.outputAsset.address
-  )
-    return;
-
-  const params = {
-    error_code: quote.error_code,
-    reason: quote.message,
-    inputAsset: { address: inputAsset.address, chainId: inputAsset.chainId, symbol: inputAsset.symbol },
-    outputAsset: { address: outputAsset.address, chainId: outputAsset.chainId, symbol: outputAsset.symbol },
-    inputAmount,
-    outputAmount,
-  };
-  lastParams = params;
-  runOnJS(analyticsTrack)(analyticsV2.event.swapsQuoteFailed, params);
-}
 
 export function useSwapInputsController({
   focusedInput,
